some styles and moved directions api
