+++ src/components/ListOfPlaces.js
@@ -1,4 +1,4 @@
-// import Directions from './Directions';
+import Directions from './Directions';
 import { Link } from 'react-router-dom';
 
 const ListOfPlaces = (props) => {
@@ -9,16 +9,14 @@ const ListOfPlaces = (props) => {
         <div>
             <ul>
                 {props.destination.map( (place) => {
-                    const name = place.name;
                     // console.log(place)
                     return(
-                        <Link to={`/directions${place.slug}`}>
-                            <li key={place.id}>
-                                {place.displayString}
-                            </li>
-                        </Link>
+                            <Link to={`/directions${place.slug}`} key={place.id}>
+                                <li>
+                                    {place.displayString}
+                                </li>
+                            </Link>
                     )
-                    // {<Directions userCoords={props.userCoords} name={name}/> }
                 })
                 }
             </ul>
